GUYS SEVERAL CHANGES DONE
First ---Added spring security basic put admin as user name and password

i have added the session validation for the employee Controller ..

so when u guys are doing take care that in mind

+++ CA-PROJECT-FINAL/src/main/java/com/leave/project/ADMIN/CONTROLLERS/AdminHollydayController.java
@@ -1,5 +1,7 @@
 package com.leave.project.ADMIN.CONTROLLERS;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -9,6 +11,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
+import com.leave.project.BEANS.UserSession;
+import com.leave.project.MODELS.Employee;
 import com.leave.project.MODELS.PublicHollyday;
 import com.leave.project.REPOSITORIES.PublicHollydayRepo;
 
@@ -20,33 +24,85 @@ public class AdminHollydayController {
 		this.phRepo = phRepo;
 	}
 	
-	@GetMapping(path="/Hollyday/view")
-	public String viewHollydays(Model model) {
+	@GetMapping(path="/admin/Hollyday/view")
+	public String viewHollydays(Model model,HttpSession session ) {
+		
+		UserSession temp= (UserSession)session.getAttribute("USER");
+		model.addAttribute("ERROR","UN-Authorized Access");
+		if(temp==null)
+				return "redirect:/*/logout";
+		Employee t=temp.getEmployee();
+		if(t == null|| !(t.getRole().getRoleName().equals("Admin"))) {
+			return "redirect:/*/logout";     ///defaulf path for logging out safely
+		}
+		
 		model.addAttribute("hollydays",phRepo.findAll());
 		return "ViewHollydays";
 	}
-	@GetMapping(path="/Hollyday/add")
-	public String getHollydayForm(Model model) {
+	@GetMapping(path="/admin/Hollyday/add")
+	public String getHollydayForm(Model model,HttpSession session) {
+		
+		UserSession temp= (UserSession)session.getAttribute("USER");
+		model.addAttribute("ERROR","UN-Authorized Access");
+		if(temp==null)
+				return "redirect:/*/logout";
+		Employee t=temp.getEmployee();
+		if(t == null|| !(t.getRole().getRoleName().equals("Admin"))) {
+			return "redirect:/*/logout";     ///defaulf path for logging out safely
+		}
+		
+		
 		model.addAttribute("hollyday",new PublicHollyday());
 		return "Hollydays";
 	}
-	@PostMapping(path="/Hollyday")
-	public String saveHollyday(Model model,PublicHollyday E) {
+	@PostMapping(path="/admin/Hollyday")
+	public String saveHollyday(Model model,PublicHollyday E,HttpSession session) {
+		
+		UserSession temp= (UserSession)session.getAttribute("USER");
+		model.addAttribute("ERROR","UN-Authorized Access");
+		if(temp==null)
+				return "redirect:/*/logout";
+		Employee t=temp.getEmployee();
+		if(t == null|| !(t.getRole().getRoleName().equals("Admin"))) {
+			return "redirect:/*/logout";     ///defaulf path for logging out safely
+		}
+		
 		phRepo.save(E);
-		return "redirect:/Hollyday/view";
+		return "redirect:/admin/Hollyday/view";
 	}
 
 	
-	 @GetMapping(path="/Hollyday/edit/{id}") 
-	 public String editHollyday(@PathVariable(name = "id") int id,Model model) {
+	 @GetMapping(path="/admin/Hollyday/edit/{id}") 
+	 public String editHollyday(@PathVariable(name = "id") int id,Model model,HttpSession session) {
+			
+			UserSession temp= (UserSession)session.getAttribute("USER");
+			model.addAttribute("ERROR","UN-Authorized Access");
+			if(temp==null)
+					return "redirect:/*/logout";
+			Employee t=temp.getEmployee();
+			if(t == null|| !(t.getRole().getRoleName().equals("Admin"))) {
+				return "redirect:/*/logout";     ///defaulf path for logging out safely
+			}
+		 
 		 model.addAttribute("hollyday",phRepo.findById(id)); 
 		 return "Hollydays";
 	 }
 	 
 	
 //for deletion simple 
-	@RequestMapping(path ="/Hollyday/delete/{id}", method = RequestMethod.GET)
-	public String deleteProduct(@PathVariable(name = "id") int id,Model model){
+	@RequestMapping(path ="/admin/Hollyday/delete/{id}", method = RequestMethod.GET)
+	public String deleteProduct(@PathVariable(name = "id") int id,Model model,HttpSession session){
+		
+		UserSession temp= (UserSession)session.getAttribute("USER");
+		model.addAttribute("ERROR","UN-Authorized Access");
+		if(temp==null)
+				return "redirect:/*/logout";
+		Employee t=temp.getEmployee();
+		if(t == null|| !(t.getRole().getRoleName().equals("Admin"))) {
+			return "redirect:/*/logout";     ///defaulf path for logging out safely
+		}
+		
+		
 		phRepo.delete(phRepo.findById(id).orElse(null)); 
 		return "redirect:/Hollyday/view";
 	}
